+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Portable Abstractions and Patterns for Python Changelog
 
+## <a name="3.1.5-3.1.6"></a> 3.1.5-3.1.6 (2021-02-26)
+
+### Bug Fixes
+
+* Fixed BooleanConverter.to_nullable_boolean
+* Fixed TypeMatcher.match_type
+
 ## <a name="3.1.3-3.1.4"></a> 3.1.3-3.1.4 (2021-01-16)
 
 ### Bug Fixes